refactor: 계산하는 로직 hook으로 분리
